requirements.txt
requests==2.31.0
pyodbc==5.0.1
python-dotenv==1.0.0
pyinstaller>=6.0.0
packaging==23.2
schedule==1.2.1
pillow>=10.0.0
PyQt6
